requirements.txt
Flask-Hashing
Pillow=10.2.0